6N11 + 6N30-1 en version 2d
