+++ src/js/exercices/6e/6N30-1.js
@@ -1,9 +1,14 @@
 import Exercice from '../Exercice.js'
 import { context } from '../../modules/context.js'
-import { calcul, choice, htmlConsigne, lettreDepuisChiffre, combinaisonListes, listeQuestionsToContenu, randint } from '../../modules/outils.js'
-import { SvgReperageSurUnAxe, LatexReperageSurUnAxe } from '../../modules/macroSvgJs.js'
+import { calcul, choice, htmlConsigne, lettreDepuisChiffre, combinaisonListes, listeQuestionsToContenu, randint, texNombre } from '../../modules/outils.js'
+import { ajouteChampTexteMathLive, setReponse } from '../../modules/gestionInteractif.js'
+import { droiteGraduee2, mathalea2d } from '../../modules/2d.js'
 
 export const titre = 'Lire l’abscisse décimale d’un point repéré par une fraction'
+export const interactifReady = true
+export const interactifType = 'mathLive'
+export const amcReady = true
+export const amcType = 'AMCOpen'
 
 /**
  * Exercice calqué sur lire abscisse fractionnaire sauf que le résultat attendu est en écriture décimale.
@@ -11,7 +16,7 @@ export const titre = 'Lire l’abscisse décimale d’un point repéré par une 
  * @author Jean-Claude Lhote
  * Référence 6N30-1
  */
-export default function LireAbscisseDecimaleBis () {
+export default function LireAbscisseDecimaleBis2d () {
   Exercice.call(this) // Héritage de la classe Exercice()
   this.consigne = 'Lire l’abscisse de chacun des points suivants et donner le résultat sous la forme d’un nombre en écriture décimale.'
   this.nbQuestions = 3
@@ -21,13 +26,14 @@ export default function LireAbscisseDecimaleBis () {
   this.spacing = 1
   this.spacingCorr = 1
   this.sup = 1
-  this.listePackages = 'tkz-euclide'
+  this.interactif = false
 
   this.nouvelleVersion = function (numeroExercice) {
     // numeroExercice est 0 pour l'exercice 1
     let typesDeQuestions
     this.listeQuestions = []
     this.listeCorrections = []
+    this.autoCorrection = []
     this.contenu = '' // Liste de questions
     this.contenuCorrection = '' // Liste de questions corrigées
     if (parseInt(this.sup) === 5) { typesDeQuestions = combinaisonListes([1, 2, 3], this.nbQuestions) } else {
@@ -36,27 +42,15 @@ export default function LireAbscisseDecimaleBis () {
         this.nbQuestions
       )
     }
-
+    const d = []
     this.contenu = htmlConsigne(this.consigne)
-    for (let i = 0,
-      abs0,
-      l1,
-      l2,
-      l3,
-      x1,
-      x2,
-      x3,
-      x11,
-      x22,
-      x33,
-      pas1,
-      pas2,
-      idUnique,
-      texte,
-      texteCorr; i < this.nbQuestions; i++) {
+    for (let i = 0, abs0, l1, l2, l3, x1, x2, x3, x11, x22, x33, xA, xB, xC, pas1, pas2, texte = '', texteCorr = '', cpt = 0; i < this.nbQuestions && cpt < 50;) {
       l1 = lettreDepuisChiffre(i * 3 + 1)
       l2 = lettreDepuisChiffre(i * 3 + 2)
       l3 = lettreDepuisChiffre(i * 3 + 3)
+      this.autoCorrection[3 * i] = { propositions: [{ statut: 4, feedback: '' }] }
+      this.autoCorrection[3 * i + 1] = { propositions: [{ statut: 4, feedback: '' }] }
+      this.autoCorrection[3 * i + 2] = { propositions: [{ statut: 4, feedback: '' }] }
       switch (typesDeQuestions[i]) {
         case 3: // Placer des demis ou des quarts sur un axe
           abs0 = 0
@@ -87,91 +81,68 @@ export default function LireAbscisseDecimaleBis () {
       x11 = randint(1, pas2 - 1)
       x22 = randint(1, pas2 - 1)
       x33 = randint(1, pas2 - 1)
-      if (context.isHtml) {
-        idUnique = `${i}_${Date.now()}`
-        this.contenu += `<div id="div_svg${numeroExercice}${idUnique}" style="width: 90%; height: 200px;  "></div>`
-        SvgReperageSurUnAxe(
-          `div_svg${numeroExercice}${idUnique}`,
-          abs0,
-          6,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11],
-            [l2, x2, x22],
-            [l3, x3, x33]
-          ],
-          [
-            [calcul(abs0 + 1 / pas1), 1, 0],
-            [calcul(abs0 + 2 / pas1), 2, 0],
-            [calcul(abs0 + 3 / pas1), 3, 0],
-            [calcul(abs0 + 4 / pas1), 4, 0],
-            [calcul(abs0 + 5 / pas1), 5, 0],
-            [calcul(abs0 + 6 / pas1), 6, 0]
-          ],
-          false
-        )
-        this.contenuCorrection += `<div id="div_svg_corr${numeroExercice}${idUnique}" style="width: 90%; height: 200px;  "></div>`
-        SvgReperageSurUnAxe(
-          `div_svg_corr${numeroExercice}${idUnique}`,
-          abs0,
-          6,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11, true],
-            [l2, x2, x22, true],
-            [l3, x3, x33, true]
-          ],
-          [
-            [calcul(abs0 + 1 / pas1), 1, 0],
-            [calcul(abs0 + 2 / pas1), 2, 0],
-            [calcul(abs0 + 3 / pas1), 3, 0],
-            [calcul(abs0 + 4 / pas1), 4, 0],
-            [calcul(abs0 + 5 / pas1), 5, 0],
-            [calcul(abs0 + 6 / pas1), 6, 0]
-          ],
-          false
-        )
+
+      xA = calcul(x1 + x11 / pas2)
+      xB = calcul(x2 + x22 / pas2)
+      xC = calcul(x3 + x33 / pas2)
+      d[2 * i] = droiteGraduee2({
+        Unite: 4,
+        Min: 0,
+        Max: 7.1,
+        axeStyle: '->',
+        pointTaille: 5,
+        pointStyle: 'x',
+        labelsPrincipaux: false,
+        thickSec: true,
+        thickSecDist: 1 / pas2,
+        labelListe: [[0, `${texNombre(abs0)}`], [1, `${texNombre(calcul(abs0 + 1 / pas1))}`]],
+        pointListe: [[xA, l1], [xB, l2], [xC, l3]]
+      })
+      d[2 * i + 1] = droiteGraduee2({
+        Unite: 4,
+        Min: 0,
+        Max: 7.1,
+        axeStyle: '->',
+        pointTaille: 5,
+        pointStyle: 'x',
+        labelsPrincipaux: false,
+        thickSec: true,
+        thickSecDist: 1 / pas2,
+        labelListe: [
+          [0, `${texNombre(abs0)}`],
+          [xA, texNombre(calcul(xA / pas1 + abs0))],
+          [xB, texNombre(calcul(xB / pas1 + abs0))],
+          [xC, texNombre(calcul(xC / pas1 + abs0))]
+        ],
+        pointListe: [[xA, l1], [xB, l2], [xC, l3]]
+
+      })
+
+      texte = mathalea2d({ xmin: -2, ymin: -1, xmax: 30, ymax: 1, pixelsParCm: 20, scale: 0.5 }, d[2 * i])
+      texteCorr = mathalea2d({ xmin: -2, ymin: -2, xmax: 30, ymax: 2, pixelsParCm: 20, scale: 0.5 }, d[2 * i + 1])
+
+      if (this.interactif && context.isHtml) {
+        setReponse(this, 3 * i, calcul(xA / pas1 + abs0))
+        setReponse(this, 3 * i + 1, calcul(xB / pas1 + abs0))
+        setReponse(this, 3 * i + 2, calcul(xC / pas1 + abs0))
+        texte += l1 + ajouteChampTexteMathLive(this, 3 * i)
+        texte += l2 + ajouteChampTexteMathLive(this, 3 * i + 1)
+        texte += l3 + ajouteChampTexteMathLive(this, 3 * i + 2)
       } else {
-        // sortie Latex
-        texte = LatexReperageSurUnAxe(
-          2,
-          abs0,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11],
-            [l2, x2, x22],
-            [l3, x3, x33]
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0]
-          ],
-          false
-        )
-        texteCorr = LatexReperageSurUnAxe(
-          2,
-          abs0,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11, true],
-            [l2, x2, x22, true],
-            [l3, x3, x33, true]
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0]
-          ],
-          false
-        )
+        if (context.isAmc) {
+          this.autoCorrection[i].enonce = texte
+          this.autoCorrection[i].propositions[0].texte = texteCorr
+        }
+      }
+      if (this.listeQuestions.indexOf(texte) === -1) {
+        // Si la question n'a jamais été posée, on en crée une autre
         this.listeQuestions.push(texte)
         this.listeCorrections.push(texteCorr)
+        i++
       }
+      cpt++
     }
-    if (!context.isHtml) { listeQuestionsToContenu(this) }
+    listeQuestionsToContenu(this)
   }
   this.besoinFormulaireNumerique = [
     'Niveau de difficulté',
